doc tools _index.md: update GENIVI project name
Whilst VSS was originally developed within the GENIVI Remote Vehicle
Interaction expert group as the GENIVI scope evolved into the wider
connected vehicle the work was continued in follow on projects with
renewed charters that reflected industry needs, such as the Common
Vehicle Interface Initiative (CVII). Update the text to reflect that
so as to not confuse newcomers.

Viva RVI, viva VSS!

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>